# openapi.yaml
openapi: 3.0.3
info:
  title: ClubOps Phase 2 API
  version: 0.1.0
  license:
    name: UNLICENSED
  description: |
    Phase 1 scope: staff PIN auth, device allowlist, register sessions (lifecycle + heartbeat + TTL), admin force sign-out, audit writes.
    Phase 2 adds: inventory items, key tags, and cleaning batches.

    Device authentication is required via headers for all non-health endpoints:
    - x-device-id
    - x-device-token

    Staff/admin authentication uses an opaque server session token in:
    - Authorization: Bearer <sessionToken>

    Heartbeat/TTL configuration (server-side):
    - REGISTER_SESSION_TTL_SECONDS (default: 900 seconds / 15 minutes)
    - REGISTER_SESSION_CLEANUP_INTERVAL_SECONDS (default: 30 seconds)
    - STAFF_SESSION_TTL_SECONDS (default: 43200 seconds / 12 hours)
servers:
  - url: /v1
...
paths:
...
  /customers:
    get:
      tags: [customers]
      summary: Search customers
      operationId: customersSearch
      security:
        - DeviceIdHeader: []
          DeviceTokenHeader: []
          StaffBearerAuth: []
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Customer search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerListResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'

    post:
      tags: [customers]
      summary: Create customer
      operationId: customersCreate
      security:
        - DeviceIdHeader: []
          DeviceTokenHeader: []
          StaffBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDTO'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /visits/active:
    get:
      tags: [visits]
      summary: Get active visit for a customer
      operationId: visitsActive
      security:
        - DeviceIdHeader: []
          DeviceTokenHeader: []
          StaffBearerAuth: []
      parameters:
        - name: customerId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Active visit or null
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitNullableResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /visits/open:
    post:
      tags: [visits]
      summary: Open a visit
      operationId: visitsOpen
      security:
        - DeviceIdHeader: []
          DeviceTokenHeader: []
          StaffBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenVisitRequest'
      responses:
        '201':
          description: Visit opened
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitDTO'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /visits/{id}/renew:
    post:
      tags: [visits]
      summary: Renew a visit
      operationId: visitsRenew
      security:
        - DeviceIdHeader: []
          DeviceTokenHeader: []
          StaffBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenewVisitRequest'
      responses:
        '200':
          description: Visit renewed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitDTO'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /visits/{id}/assign:
    post:
      tags: [visits]
      summary: Assign inventory item to visit
      operationId: visitsAssign
      security:
        - DeviceIdHeader: []
          DeviceTokenHeader: []
          StaffBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignVisitRequest'
      responses:
        '200':
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitAssignmentDTO'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /visits/{id}/close:
    post:
      tags: [visits]
      summary: Close a visit
      operationId: visitsClose
      security:
        - DeviceIdHeader: []
          DeviceTokenHeader: []
          StaffBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Visit closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitDTO'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    DeviceIdHeader:
      type: apiKey
      in: header
      name: x-device-id
    DeviceTokenHeader:
      type: apiKey
      in: header
      name: x-device-token
    StaffBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
          enum:
            - DEVICE_UNAUTHORIZED
            - DEVICE_DISABLED
            - DEVICE_NOT_FOUND
            - STAFF_UNAUTHORIZED
            - STAFF_SESSION_EXPIRED
            - STAFF_DISABLED
            - FORBIDDEN
            - VALIDATION_ERROR
            - RATE_LIMITED
            - REGISTER_SESSION_NOT_FOUND
            - REGISTER_SESSION_NOT_ACTIVE
            - REGISTER_ACTIVE_CONFLICT
            - DEVICE_ACTIVE_CONFLICT
            - INVENTORY_NOT_FOUND
            - INVALID_STATUS_TRANSITION
            - KEY_TAG_NOT_FOUND
            - KEY_TAG_DISABLED
            - KEY_TAG_ASSIGNMENT_CONFLICT
            - CUSTOMER_NOT_FOUND
            - VISIT_NOT_FOUND
            - VISIT_ALREADY_ACTIVE
            - VISIT_MAX_DURATION_EXCEEDED
            - INVENTORY_UNAVAILABLE_FOR_ASSIGNMENT
            - AGREEMENT_ALREADY_CAPTURED
...
    VisitStatus:
      type: string
      enum: [ACTIVE, CLOSED]
...
    CustomerDTO:
      type: object
      required: [id, firstName, lastName, displayName, phone, email, createdAt]
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        displayName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    CustomerListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CustomerDTO'
    CreateCustomerRequest:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        displayName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
    VisitDTO:
      type: object
      required:
        [
          id,
          customerId,
          status,
          startedAt,
          plannedEndAt,
          closedAt,
          initialDurationMinutes,
          maxTotalDurationMinutes,
          renewalTotalMinutes
        ]
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/VisitStatus'
        startedAt:
          type: string
          format: date-time
        plannedEndAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
          nullable: true
        initialDurationMinutes:
          type: integer
        maxTotalDurationMinutes:
          type: integer
        renewalTotalMinutes:
          type: integer
    VisitNullableResponse:
      type: object
      required: [visit]
      properties:
        visit:
          $ref: '#/components/schemas/VisitDTO'
          nullable: true
    OpenVisitRequest:
      type: object
      required: [customerId]
      properties:
        customerId:
          type: string
          format: uuid
    RenewVisitRequest:
      type: object
      required: [durationMinutes]
      properties:
        durationMinutes:
          type: integer
          enum: [120, 360]
    AssignVisitRequest:
      type: object
      required: [inventoryItemId]
      properties:
        inventoryItemId:
          type: string
          format: uuid
    VisitAssignmentDTO:
      type: object
      required: [visitId, inventoryItemId, assignedAt, releasedAt]
      properties:
        visitId:
          type: string
          format: uuid
        inventoryItemId:
          type: string
          format: uuid
        assignedAt:
          type: string
          format: date-time
        releasedAt:
          type: string
          format: date-time
          nullable: true
...